tests/requirements-test.txt
# =====================================================================
# MUTT v2.3 Test Dependencies
# =====================================================================
# Install with: pip install -r requirements-test.txt
# =====================================================================

# Core testing framework
pytest
pytest-cov==4.1.0          # Coverage reporting
pytest-mock==3.15.1        # Mocking helpers
pytest-xdist==3.8.0        # Parallel test execution

# Testing utilities
coverage==7.10.7            # Coverage measurement
mock==5.1.0                # Additional mocking support

# Code quality (optional)
pytest-flake8==1.3.0       # Linting during tests
pytest-pylint==0.21.0      # Code quality checks

# Performance testing (optional)
pytest-benchmark==5.2.3    # Performance benchmarking

# Test reporting
pytest-html==4.1.1         # HTML test reports
pytest-json-report==1.5.0  # JSON test reports

# Main application dependencies (from requirements.txt)
Flask==3.1.2
redis==5.0.1
psycopg2-binary==2.9.11
hvac==2.0.0
requests==2.32.5
requests-mock==1.12.1
prometheus-client==0.18.0
prometheus-flask-exporter==0.23.0

# =====================================================================
# Installation:
# =====================================================================
# 1. Create/activate virtual environment:
#    python3 -m venv venv
#    source venv/bin/activate
#
# 2. Install test dependencies:
#    pip install -r requirements-test.txt
#
# 3. Run tests:
#    pytest tests/ -v
#
# 4. Run with coverage:
#    pytest tests/ --cov=. --cov-report=html
#
# 5. View coverage report:
#    open htmlcov/index.html
# =====================================================================
